finish p5-task2

+++ include/os/string.h
@@ -15,6 +15,6 @@ int strlen(const char *src);
 
 // [p4] safety copy
 void copyin(uint8_t *ker_dst, uint8_t *usr_src, uint32_t len);
-void copyout(uint32_t *ker_src, uint8_t *usr_dst, uint32_t len);
+void copyout(uint8_t *ker_src, uint8_t *usr_dst, uint32_t len);
 
 #endif
